+++ projects/demo_app/src/app/app.module.ts
@@ -1,26 +1,45 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import * as PlotlyJS from 'plotly.js-dist';
-import { PlotlyModule } from '../../../plotly/src/lib/plotly.module';
-import { PlotlyViaCDNModule } from 'projects/plotly/src/lib/plotly-via-cdn.module';
-
-
-
+import { RouterModule } from '@angular/router';
+import PlotlyJS from 'plotly.js-dist';
+import { PlotlyModule } from 'angular-plotly.js';
 import { AppComponent } from './app.component';
-import { PlotlyViaWindowModule } from 'projects/plotly/src/lib/plotly-via-window.module';
-
+import { routes } from './app.routes';
+import { DashboardComponent } from './pages/dashboard/dashboard.component';
+import { EventsComponent } from './pages/events/events.component';
+import { ExamplesIndexComponent } from './pages/examples-index/examples-index.component';
+import { GettingStartedComponent } from './pages/getting-started/getting-started.component';
+import { HomeComponent } from './pages/home/home.component';
+import { LifecycleComponent } from './pages/lifecycle/lifecycle.component';
+import { LoadingComponent } from './pages/loading/loading.component';
+import { ReactiveUpdatesComponent } from './pages/reactive-updates/reactive-updates.component';
+import { ChartCardComponent } from './shared/chart-card.component';
+import { CodeBlockComponent } from './shared/code-block.component';
+import { ExamplePageComponent } from './shared/example-page.component';
 
 @NgModule({
   declarations: [
-    AppComponent
+    AppComponent,
+    ChartCardComponent,
+    CodeBlockComponent,
+    DashboardComponent,
+    EventsComponent,
+    ExamplePageComponent,
+    ExamplesIndexComponent,
+    GettingStartedComponent,
+    HomeComponent,
+    LifecycleComponent,
+    LoadingComponent,
+    ReactiveUpdatesComponent,
   ],
   imports: [
     BrowserModule,
-    // PlotlyModule.forRoot(PlotlyJS),
-    // PlotlyViaCDNModule.forRoot({version: '3.0.1'}),
-    PlotlyViaWindowModule,
+    RouterModule.forRoot(routes, {
+      anchorScrolling: 'enabled',
+      scrollPositionRestoration: 'enabled',
+    }),
+    PlotlyModule.forRoot(PlotlyJS),
   ],
-  providers: [],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
 export class AppModule { }